requirements.txt
sqlalchemy
fastapi
mysql-connector-python
uvicorn[standard]